Add an experimental feature to Gamma-correct LED
There was previously an issue with intensities of the LED's not changing
noticably enough. An atempt has been made here to adjust this issue, by
mapping the linear intensities to Gamma corrected ones.

As there are only 8 values per channel, this was done via a lookup array.

This features is expterimental as of this time, and needs to be tested.