Add base input handling, and some debug functions
Add an untested input handler, which parses the physical inputs from the
rotary encoder. This includes a software debounce function.

Also started to implement some of the view controllers, which will
govern how the interface responds to user input. Currently, they are
untested and only a couple are implemented.

A possible problem is how the hand-off is to take place between the
different controllers, without the previous inputs being interpreted
as new ones. Probably, another tracking variable can be used, where we
record if the input has been used for anything.

Furthermore, I have added some debuging functions, so that the rotary
encoder and the LED panels can be tested. These are turned on via
switches in the .h file, which means that the program needs to be
recompiled for these to take effect.